requirements.txt
Flask
numpy
Pillow
scikit-learn
joblib